refactor: update component styles file names to follow naming convention
This commit refactors the names of component styles files across the project
to adhere to a new naming convention. Previously, the files used the pattern
"ComponentNameStyles.js". With the changes, the files have been renamed to
"ComponentName.styles.js" to align with a more standardized and descriptive
naming convention.

The renaming affects multiple files throughout the project, updating the
file names to reflect the new convention.

This refactoring improves the organization and consistency of the codebase,
making it easier to locate and identify the associated styles for each
component. The changes enhance codebase maintainability and align with best
practices for naming conventions.

No functional changes have been made to the code itself, only the file names
have been updated.
