+++ california-dev/src/components/ui/FloatingActionButton/FloatingActionButton.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 
-import { FabContainer, FabButton, FabMenu } from "./FabStyles";
+import { FabContainer, FabButton, FabMenu } from "./Fab.styles";
 import { CogIcon } from "assets/images/icons/CogIcon";
 export const FloatingActionButton = ({ toggleTheme }) => {
   const [expanded, setExpanded] = useState(false);
